Refactor user-service for modularity
