+++ backend/user-service/routes/user.js
@@ -1,6 +1,6 @@
 import dotenv from "dotenv";
 import express from "express";
-import { updateUserName } from '../database.js';
+import { updateUserName } from '../db/queries/userQueries.js';
 
 dotenv.config();
 const router = express.Router();
